Ontologies

The Ontology Lookup Service (OLS) is designed to leverage the power of multiple ontologies, providing a structured and semantic framework for organizing, interpreting, and querying data across various domains.

The OLS provides access to a wide range of ontologies, including the following:

ChEBI

GO

NCBITaxon

PR

UBERON

EFO

The OLS also provides access to the following cross-domain ontologies:

OBO

EDAM

And many more. The OLS is designed to be extensible, allowing users to add their own ontologies to the system.

Listing ontologies

To list all available ontologies, use the following command:

list_ontologies()

This will return a list of all available ontologies in the OLS.

Note, this will return Ontology objects, which contain information about the ontology, such as the name, title, description, and version. If you wish to only get the ids of the ontologies, you can use the following command:

list_ontology_ids()

Getting an ontology

To get information about a specific ontology, use the following command:

get_ontology("CHEBI")

References

OntologyLookup.Ontologies.get_ontology — Method

get_ontology(ontology_id::String)::Union{Ontology,Missing}

Fetches an ontology from the OLS (Ontology Lookup Service) based on the given ontology_id. To get a list of available ontologies, use the list_ontologies() function.

Returns

If the ontology is found, returns an Ontology object containing information about the ontology.

If the ontology is not found or an error occurs during the fetch, returns missing.

See also

list_ontologies()

list_ontologies_ids()

OntologyLookup.Ontologies.list_ontologies — Method

list_ontologies()::Union{Vector{Ontology},Missing}

Fetches a list of ontologies from the OLS (Ontology Lookup Service) API.

Returns

If successful, returns a vector of Ontology objects representing the fetched ontologies.

If an error occurs during the API request, returns missing.

See also

list_ontologies_ids()

get_ontology()

OntologyLookup.Ontologies.list_ontologies_ids — Method

list_ontologies_ids()::Union{Vector{String},Missing}

Returns a vector of ontology IDs.

Retrieve the list of available ontologies, and returns a vector of their IDs. If an error occurs during the fetch, the function returns missing.

See also

list_ontologies()

get_ontology()